+++ src/PythonPlugin/PythonConsoleView.cpp
@@ -225,7 +225,7 @@ PythonConsoleViewImpl::PythonConsoleViewImpl(PythonConsoleView* self)
     python::object keyword = python::module::import("keyword");
 #ifdef CNOID_USE_PYBIND11
     pybind11::list kwlist = pybind11::cast<pybind11::list>(keyword.attr("kwlist"));
-    for(int i = 0; i < pybind11::len(kwlist); ++i){
+    for(size_t i = 0; i < pybind11::len(kwlist); ++i){
         keywords.push_back(pybind11::cast<string>(kwlist[i]));
     }
 #else
@@ -371,7 +371,7 @@ std::vector<string> PythonConsoleViewImpl::getMemberNames(python::object& module
     python::list memberNames = python::extract<python::list>(python::object(h));
 #endif
     std::vector<string> retNames;
-    for(int i=0; i < python::len(memberNames); ++i){
+    for(size_t i=0; i < python::len(memberNames); ++i){
 #ifdef CNOID_USE_PYBIND11
         if(!strstr(string(memberNames[i].cast<string>()).c_str(), "__" )){
             retNames.push_back(string(memberNames[i].cast<string>()));
@@ -403,10 +403,10 @@ void PythonConsoleViewImpl::tabComplete()
     beforeCursorString = beforeCursorString.substr(0,cursor.columnNumber()-inputColumnOffset);
     QString afterCursorString = getInputString();
     afterCursorString.remove(0, cursor.columnNumber()-inputColumnOffset);
-    int maxSplitIdx = 0;
+    size_t maxSplitIdx = 0;
     for(std::vector<string>::iterator it = splitStringVec.begin(); it != splitStringVec.end();  ++it){
-        int splitIdx = beforeCursorString.find_last_of(*it);
-        maxSplitIdx = std::max(splitIdx == string::npos ? 0 : splitIdx+1,maxSplitIdx);
+        size_t splitIdx = beforeCursorString.find_last_of(*it);
+        maxSplitIdx = std::max(splitIdx == string::npos ? 0 : splitIdx + 1, maxSplitIdx);
     }
     string lastWord = beforeCursorString.substr(maxSplitIdx);
     beforeCursorString = beforeCursorString.substr(0,maxSplitIdx);
@@ -431,7 +431,7 @@ void PythonConsoleViewImpl::tabComplete()
 
     std::vector<string> completions;
     unsigned long int maxLength = std::numeric_limits<long>::max();
-    for(int i=0; i < memberNames.size(); ++i){
+    for(size_t i=0; i < memberNames.size(); ++i){
         if(memberNames[i].substr(0,lastDottedString.size()) == lastDottedString){
             completions.push_back(memberNames[i]);
             maxLength = std::min((unsigned long int)memberNames[i].size(),maxLength);
@@ -449,9 +449,9 @@ void PythonConsoleViewImpl::tabComplete()
     if(completions.size() != 0){
         // max common string among completions
         std::string maxCommonStr = lastDottedString;
-        for(int i=maxCommonStr.size(); i < maxLength; ++i){
+        for(size_t i=maxCommonStr.size(); i < maxLength; ++i){
             bool commomFlg = true;
-            for(int j=1; j < completions.size(); ++j){
+            for(size_t j=1; j < completions.size(); ++j){
                 if(completions[0].at(i) != completions[j].at(i)){
                     commomFlg = false;
                     break;
@@ -459,7 +459,7 @@ void PythonConsoleViewImpl::tabComplete()
             }
             if( commomFlg ){
                 maxCommonStr.push_back(completions[0].at(i));
-            }else{
+            } else {
                 break;
             }
         }
@@ -474,7 +474,7 @@ void PythonConsoleViewImpl::tabComplete()
             put("\n"); // This must be done after getInputString().
 
             string str = "";
-            for(int i=0; i < completions.size(); ++i){
+            for(size_t i=0; i < completions.size(); ++i){
                 str.append(beforeLastDotStr);
                 str.append(completions[i]);
                 str.append("     ");
@@ -489,7 +489,7 @@ void PythonConsoleViewImpl::tabComplete()
         str.append(maxCommonStr);
         str.append(afterCursorString.toStdString());
         setInputString(QString(str.c_str()));
-        for(int i=0; i < afterCursorString.toStdString().size(); ++i){
+        for(size_t i=0; i < afterCursorString.toStdString().size(); ++i){
             moveCursor(QTextCursor::Left);
         }
     }
